Replace old evaluation with new sdmetrics
